+++ tfds/tf_multi_prio_queue.hpp
@@ -45,7 +45,7 @@ public:
     }
 
     // O(1)
-    int length() const {
+    size_t length() const {
         return tree.size();
     }
